// src/app/main/interfaces/costumer.interface.ts
export interface Customer {
  country: string | null;
  id: string | null;
  firstName: string | null;
  lastName: string | null;
  birthDate: string | null;
  gender: string | null;
  cellPhone: string | null;
  homePhone: string | null;
  addressHome: string | null;
  profession: string | null;
  incomes: string | null;
}